server/requirements.txt
django>=1.9.0
django-rest-auth==0.9.2
djangorestframework>=3.7.0
django-allauth>=0.24.1
six==1.9.0
